% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFun.r
\name{myFun}
\alias{myFun}
\title{myFun}
\usage{
myFun(arg1, arg2 = 1)
}
\arguments{
\item{arg1}{A number}

\item{arg2}{A number with default value}
}
\value{
The sum of \code{arg1} and \code{arg2}
}
\description{
This function sums up two numbers.
}
\details{
\code{data} contains the transaction data. The data set must contain a
            column labeled "Customer" that allows unique customer identification
            and a column labeled "TransDate", indicating the purchase date.
            The column "PurchAmount" specifies the total spending per purchase.
}
\examples{
myFun(1, 1)
myFun(10, 1)

}
